Improve packaging
