+++ src/cmd/csi/provisioner/builder.go
@@ -1,4 +1,4 @@
-package csi_server
+package provisioner
 
 import (
 	"path/filepath"
@@ -6,8 +6,9 @@ import (
 	"github.com/Dynatrace/dynatrace-operator/src/cmd/config"
 	cmdManager "github.com/Dynatrace/dynatrace-operator/src/cmd/manager"
 	dtcsi "github.com/Dynatrace/dynatrace-operator/src/controllers/csi"
-	csidriver "github.com/Dynatrace/dynatrace-operator/src/controllers/csi/driver"
+	csigc "github.com/Dynatrace/dynatrace-operator/src/controllers/csi/gc"
 	"github.com/Dynatrace/dynatrace-operator/src/controllers/csi/metadata"
+	csiprovisioner "github.com/Dynatrace/dynatrace-operator/src/controllers/csi/provisioner"
 	"github.com/pkg/errors"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
@@ -15,12 +16,10 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
-const use = "csi-server"
+const use = "csi-provisioner"
 
 var (
-	nodeId       = ""
 	probeAddress = ""
-	endpoint     = ""
 )
 
 type CommandBuilder struct {
@@ -31,7 +30,7 @@ type CommandBuilder struct {
 	csiOptions      *dtcsi.CsiOptions
 }
 
-func NewCsiServerCommandBuilder() CommandBuilder {
+func NewCsiProvisionerCommandBuilder() CommandBuilder {
 	return CommandBuilder{}
 }
 
@@ -63,9 +62,7 @@ func (builder CommandBuilder) setFilesystem(filesystem afero.Fs) CommandBuilder 
 func (builder CommandBuilder) getCsiOptions() dtcsi.CsiOptions {
 	if builder.csiOptions == nil {
 		builder.csiOptions = &dtcsi.CsiOptions{
-			NodeId:   nodeId,
-			Endpoint: endpoint,
-			RootDir:  dtcsi.DataPath,
+			RootDir: dtcsi.DataPath,
 		}
 	}
 
@@ -100,9 +97,7 @@ func (builder CommandBuilder) Build() *cobra.Command {
 }
 
 func addFlags(cmd *cobra.Command) {
-	cmd.PersistentFlags().StringVar(&nodeId, "node-id", "", "node id")
-	cmd.PersistentFlags().StringVar(&endpoint, "endpoint", "unix:///tmp/csi.sock", "CSI endpoint")
-	cmd.PersistentFlags().StringVar(&probeAddress, "health-probe-bind-address", ":10080", "The address the probe endpoint binds to.")
+	cmd.PersistentFlags().StringVar(&probeAddress, "health-probe-bind-address", ":10090", "The address the probe endpoint binds to.")
 }
 
 func (builder CommandBuilder) buildRun() func(*cobra.Command, []string) error {
@@ -130,12 +125,12 @@ func (builder CommandBuilder) buildRun() func(*cobra.Command, []string) error {
 			return err
 		}
 
-		err = metadata.CorrectMetadata(csiManager.GetClient(), access)
+		err = csiprovisioner.NewOneAgentProvisioner(csiManager, builder.getCsiOptions(), access).SetupWithManager(csiManager)
 		if err != nil {
 			return err
 		}
 
-		err = csidriver.NewServer(csiManager.GetClient(), builder.getCsiOptions(), access).SetupWithManager(csiManager)
+		err = csigc.NewCSIGarbageCollector(csiManager.GetClient(), builder.getCsiOptions(), access).SetupWithManager(csiManager)
 		if err != nil {
 			return err
 		}
