+++ src/cmd/csi/provisioner/command_test.go
@@ -1,4 +1,4 @@
-package csi_server
+package provisioner
 
 import (
 	"testing"
