+++ src/cmd/csi/provisioner/manager_test.go
@@ -1,4 +1,4 @@
-package csi_server
+package provisioner
 
 import (
 	"testing"
