+++ src/cmd/csi/server/builder_test.go
@@ -1,4 +1,4 @@
-package csi_server
+package server
 
 import (
 	"io/fs"
