+++ src/cmd/csi/server/command_test.go
@@ -1,4 +1,4 @@
-package csi_server
+package server
 
 import (
 	"testing"
